+++ tkviews/core/__init__.py
@@ -1,4 +1,3 @@
 """Core package"""
 
-from .node import TkNode
 from .rendering import TkRenderingContext, render_attribute