+++ tkviews/core/rendering.py
@@ -3,7 +3,7 @@
 from tkinter import Widget
 from typing import Tuple, Any
 
-from pyviews.core import InheritedDict, Node, XmlAttr, Setter
+from pyviews.core import Node, XmlAttr, Setter
 from pyviews.expression import is_expression, parse_expression, Expression, execute
 from pyviews.pipes import get_setter
 from pyviews.rendering import RenderingContext
@@ -20,15 +20,6 @@ class TkRenderingContext(RenderingContext):
     @master.setter
     def master(self, value: Widget):
         self['master'] = value
-
-    @property
-    def node_styles(self) -> InheritedDict:
-        """node styles"""
-        return self['node_styles']
-
-    @node_styles.setter
-    def node_styles(self, value: InheritedDict):
-        self['node_styles'] = value
 
 
 def render_attribute(node: Node, xml_attr: XmlAttr) -> Tuple[Setter, Any]: